+++ src/elements/conditions/assets/HasAltConditionRule.php
@@ -1,0 +1,54 @@
+<?php
+
+namespace craft\elements\conditions\assets;
+
+use Craft;
+use craft\base\conditions\BaseLightswitchConditionRule;
+use craft\base\ElementInterface;
+use craft\elements\Asset;
+use craft\elements\conditions\ElementConditionRuleInterface;
+use craft\elements\db\AssetQuery;
+use craft\elements\db\ElementQueryInterface;
+
+/**
+ * “Has alternative text” condition rule for assets.
+ *
+ * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
+ * @since 4.0.0
+ */
+class HasAltConditionRule extends BaseLightswitchConditionRule implements ElementConditionRuleInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function getLabel(): string
+    {
+        return Craft::t('app', 'Has alternative text');
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getExclusiveQueryParams(): array
+    {
+        return ['hasAlt'];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function modifyQuery(ElementQueryInterface $query): void
+    {
+        /** @var AssetQuery $query */
+        $query->hasAlt($this->value);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function matchElement(ElementInterface $element): bool
+    {
+        /** @var Asset $element */
+        return $this->matchValue($element->alt !== null);
+    }
+}